removed warnings

+++ src/layer/TiledObjectLayer.cpp
@@ -12,8 +12,8 @@ namespace phantom{
         _tilesX = x;
         _tilesY = y;
 
-        width = x * tileSize;
-        height = y * tileSize;
+        width = static_cast<float>(x * tileSize);
+        height = static_cast<float>(y * tileSize);
 
         this->width = static_cast<float>(x * _tilesX);
         this->height = static_cast<float>(y* _tilesY);
@@ -28,8 +28,8 @@ namespace phantom{
         }
     }
     Tile* TiledObjectLayer::getTileAt(Eigen::Vector3f position) throw(...){
-        int x = max<float>(0, min<float>(_tilesX -1, floorf(position.x() / _tileSize)));
-        int y = max<float>(0, min<float>(_tilesY -1, floorf(position.y() / _tileSize)));
+        int x = static_cast<int>(max<float>(0, min<float>(static_cast<int>(_tilesX) - 1.0f, floorf(position.x() / _tileSize))));
+        int y = static_cast<int>(max<float>(0, min<float>(static_cast<int>(_tilesY) - 1.0f, floorf(position.y() / _tileSize))));
         if(_tileList == 0){
             throw PhantomException("Tiles do not exist. Did you call TiledObjectLayer::createTile()?");
